update docs and another test
